rename components folders
